changes to clean up example
