+++ test/examples/15-SpdlogDynamicConfiguration/run.sh.in
@@ -42,9 +42,9 @@ SDPID=$!
 # sleep for a second to give the service directory a chance to start up
 sleep 1
 
-./SpdlogConfigurationPublisherTest &
+./SpdlogDynamicConfiguration &
 PUBID=$!
 
-sleep 30
+sleep 25
 
 cleanup
